CRITICAL FIX: Resolve FlashInfer data type dispatch error
Fix "apply_rope_with_cos_sin_cache_inplace failed to dispatch data type Float" error
by ensuring all tensors use FlashInfer-compatible data types (fp16/bf16, not float32).

Changes:
- model_hub/deepseek.py: Change RoPE cache from float32 to model dtype (fp16)
- model_hub/llama.py: Use model dtype instead of inv_freq.dtype for RoPE cache
- model_hub/qwen.py: Use model dtype instead of inv_freq.dtype for RoPE cache
- attn_hub/mla_attn.py: Match cos_sin_cache dtype to query/key tensors

Root cause: FlashInfer only supports fp16/bf16 data types for RoPE operations,
but code incorrectly used float32 based on outdated/wrong documentation.

Previous fixes:
- ✅ Fixed position_ids tensor shape (2D → 1D)
- ✅ Fixed data type compatibility (float32 → fp16/bf16)

+++ attn_hub/mla_attn.py
@@ -46,8 +46,8 @@ def apply_decoupled_rope(query_states, key_states, position_ids, rope_dim, cos_c
         # 如果已经是合并格式
         cos_sin_cache = cos_cache
     
-    # 确保在正确设备上，但保持float32数据类型（FlashInfer要求）
-    cos_sin_cache = cos_sin_cache.to(device=q_rope.device, dtype=torch.float32)
+    # 确保在正确设备上，并使用与query/key张量相同的数据类型（FlashInfer要求fp16/bf16）
+    cos_sin_cache = cos_sin_cache.to(device=q_rope.device, dtype=q_rope.dtype)
     
     # FlashInfer RoPE API需要正确的position_ids格式
     # 将二维position_ids转换为一维格式，这是FlashInfer期望的格式
